Code extraction to subscribers and handler
