+++ src/Resources/config/services/event_subscriber.xml
@@ -4,11 +4,15 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
 
-        <service id="bitbag_sylius_agreement_plugin.event.agreement_checked_subscriber"
-                 class="BitBag\SyliusAgreementPlugin\EventSubscriber\AgreementCheckedSubscriber">
-            <argument type="service" id="bitbag_sylius_agreement_plugin.repository.agreement_history"/>
-            <argument type="service" id="bitbag_sylius_agreement_plugin.resolver.agreement"/>
-            <argument type="service" id="bitbag_sylius_agreement_plugin.resolver.composite_agreement_history"/>
+        <service id="bitbag_sylius_agreement_plugin.event.user_registration_agreement_subscriber"
+                 class="BitBag\SyliusAgreementPlugin\EventSubscriber\UserRegistrationAgreementSubscriber">
+            <argument type="service" id="bitbag_sylius_agreement_plugin.handler.agreement_handler"/>
+            <tag name="kernel.event_subscriber" />
+        </service>
+
+        <service id="bitbag_sylius_agreement_plugin.event.agreement_subscriber"
+                 class="BitBag\SyliusAgreementPlugin\EventSubscriber\AgreementSubscriber">
+            <argument type="service" id="bitbag_sylius_agreement_plugin.handler.agreement_handler"/>
             <tag name="kernel.event_subscriber" />
         </service>
